+++ matching-eval-client/src/test/java/de/uni_mannheim/informatik/dws/melt/matching_eval_client/BuiltInTracksTest.java
@@ -22,9 +22,8 @@ class BuiltInTracksTest {
         assertNotNull(result);
         assertTrue(result.size() > 3);
 
-        for(String s : result){
-            System.out.println(s);
-        }
+        assertTrue(result.contains("instancematching"));
+        assertFalse(result.contains("instancematching-geolink-cruise"));
     }
 
 }